refactor(identity): consolidate TOTP RFC 6238 logic into TotpServiceBase
DefaultTotpService (Granit.OpenIddict) and TotpService (Granit.Identity.Local.AspNetIdentity)
shared a near-identical HMAC-SHA1 + Base32 implementation. Two copies of crypto-sensitive
code mean drift risk and double the security-review surface. Extract:

- Granit.Identity.Local/Internal/Base32.cs (internal helper)
- Granit.Identity.Local/Services/TotpServiceBase.cs (public abstract)

Concrete services now only supply their issuer (constant "Granit" vs IdentityOptions.Tokens.AuthenticatorIssuer).

Net: -117 LOC, one canonical source of truth for TOTP.

Drive-by fix: TotpService.cs Base32 switch arms used the raw input char instead of
the upper-cased one when computing offsets — silently incorrect on lowercase Base32 input.
The OpenIddict version was already correct; consolidation adopts that behavior.
